Bug fix 正確的使用剩餘金錢，讀取正確的槓桿

+++ exchange-list/bybit/bybitResult.go
@@ -183,7 +183,7 @@ type GetLeverageResult struct {
 }
 
 type LeverageItem struct {
-	Leverage int `json:"leverage"`
+	Leverage float64 `json:"leverage"`
 }
 
 type PositionV1 struct {
